@@ -3,19 +3,17 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@reach/router": "github:reach/router",
     "contentful": "^7.10.0",
     "embla-carousel": "^4.0.6",
     "ini": ">=1.3.6",
     "node-notifier": ">=8.0.1",
-    "node-sass": "^4.14.1",
+    "sass": "^1.69.5",
     "nvm": "0.0.4",
-    "reach-router": "0.0.0",
-    "react": "^17.0.0",
-    "react-dom": "^17.0.0",
-    "react-hook-form": "^6.10.0",
-    "react-router-dom": "5.2.0",
-    "react-scripts": "3.4.4",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-hook-form": "^7.48.0",
+    "react-router-dom": "^6.20.0",
+    "react-scripts": "5.0.1",
     "slugify": "^1.3.4"
   },
   "scripts": {
